+++ Syncro/src/uk/me/grambo/syncro/RemoteFileHandler.java
@@ -4,7 +4,7 @@ public interface RemoteFileHandler {
 	public class RemoteFileData {
 		public String Filename;
 		public int FolderId;
-		public int Size;
+		public long Size;
 	}
 	public void HandleRemoteFile(RemoteFileData inoFile);
 }
